+++ src/components/friends/sync/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v4.0.0-beta.139.dsl-speedup-2 (2024-10-03)
+
+#### :house: Internal
+
+* Performance improvements
+
 ## v4.0.0-beta.138.dsl-speedup (2024-10-01)
 
 #### :house: Internal