+++ src/core/component/render/CHANGELOG.md
@@ -15,6 +15,12 @@ Changelog
 
 * Fixed incorrect `shapeFlag` on a functional vnode
 
+## v4.0.0-beta.139.dsl-speedup-2 (2024-10-03)
+
+#### :rocket: New Feature
+
+* Added a new default prop `getPassedProps`, which allows identifying which props were passed through the template
+
 ## v4.0.0-beta.107 (2024-07-10)
 
 #### :bug: Bug Fix